% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickUDS.R
\name{democracy_model}
\alias{democracy_model}
\title{Produce a UD model from democracy data}
\usage{
democracy_model(..., identifiers = c("extended_country_name", "year", "GWn",
  "cown", "uds_2010_mean", "uds_2011_mean", "uds_2014_mean", "in_GW_system"),
  verbose = TRUE, technical = list(NCYCLES = 2500))
}
\arguments{
\item{...}{\link{democracy} variables to use for the model. Can be bare column names
or strings, or a \link[dplyr:select]{dplyr::select} expression.}

\item{identifiers}{Identifier columns. Can be any combination of columns in
the \link{democracy} dataset. Defaults to \code{c("extended_country_name",
  "year", "GWn", "cown", "uds_2010_mean", "uds_2011_mean", "uds_2014_mean")}.}

\item{verbose}{Passed to \link[mirt:mirt]{mirt::mirt}; whether to print a running commentary.
Default is \code{TRUE}.}

\item{technical}{Passed to \link[mirt:mirt]{mirt::mirt}. Defaults to \code{list(NCYCLES =
2500)} to ensure that extende models converge.}
}
\value{
a \code{\link{SingleGroupClass-class}} model of latent democracy
scores suitable for use by \code{\link{democracy_scores}}.
}
\description{
This function is a simple wrapper for \link[mirt:mirt]{mirt::mirt} that automates the process
of selecting the relevant columns in the \link{democracy} dataset, preparing the
data for use with \link[mirt:mirt]{mirt::mirt}, and packaging the result in a form useful for
\link{democracy_scores}. More fine-grained control can be achieved by using
\code{\link{mirt}} directly; see the Vignette on replicating and extending
the UD scores.
}
\examples{
\donttest{
replication_2011_model <- democracy_model(dplyr::matches("pmm"), verbose = FALSE)
replication_2011_model
summary(replication_2011_model)
}
}
