% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickUDS.R
\name{democracy_scores}
\alias{democracy_scores}
\title{A convenience function for producing UD scores from a UD model.}
\usage{
democracy_scores(model, ...)
}
\arguments{
\item{model}{a \code{\link{SingleGroupClass-class}} model produced by
\code{\link{democracy_model}}.}

\item{...}{Other parameters passed on to \code{\link{fscores}}.}
}
\value{
A data frame with latent variable democracy scores (posterior means)
  for all country-years in the data, with standard errors and 95% confidence
  intervals.
}
\description{
This function is a simple wrapper for \code{fscores(model, full.scores =
TRUE, full.scores.SE = TRUE, ...)} that returns scores in a tidy data frame
instead of a matrix. More fine-grained control can be achieved by using
\code{\link{fscores}} directly.
}
\examples{
# Replicate the official UDS scores (2011 release)
library(reshape2)
library(dplyr)
data <- prepare_data(democracy)
data <- melt(data, measure.vars = names(data)[grep("pmm",names(data))], na.rm = TRUE)
data <- data \%>\% group_by(country_name,year) \%>\% mutate(num_measures = n())
data <- dcast(data, ... ~ variable)
data <- data \%>\% arrange(country_name,year)
replication_2011_model <- democracy_model(data, columns = names(data)[grep("pmm",names(data))])
replication_2011_scores <- democracy_scores(replication_2011_model)
replication_2011_scores <- bind_cols(data \%>\%
                                       select(country_name, GWn, year,in_system,num_measures),
                                                     replication_2011_scores)
}

