% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickUDS.R
\name{match_to_uds}
\alias{match_to_uds}
\title{Match generated scores to the level of an official UD release}
\usage{
match_to_uds(data, release = 2014)
}
\arguments{
\item{data}{A dataset containing an extension of the UD scores, from the
output of \link{democracy_scores}. It must contain a \code{country_name}
column, a \code{year} column, \code{z1} column, a \code{pct025} column, and
a \code{pct975} column.}

\item{release}{The year of a Unified Democracy Scores official release (see
the available releases in their data page at
\url{http://www.unified-democracy-scores.org/uds.html} and
\url{http://www.unified-democracy-scores.org/archive.html}). Default is
2014.}
}
\value{
The dataset with three columns appended: \code{adj.z1} (the mean
  democracy score, adjusted to match the UD release), \code{adj.pct025} and
  \code{adj.pct975} (the adjusted confidence bounds).
}
\description{
Match generated scores to the level of an official UD release
}
\examples{
\donttest{
library(dplyr)
indexes <- c("arat_pmm","blm","bmr_democracy","bollen_pmm",
                            "doorenspleet","eiu","freedomhouse",
                            "gwf","hadenius_pmm","lied","mainwaring",
                            "munck_pmm","pacl","polity2","polyarchy_pmm","prc",
                            "svolik","ulfelder","vanhanen_democratization",
                            "wahman_teorell_hadenius")
data <- prepare_democracy(indexes)
extended_model <- democracy_model(data,indexes, verbose=TRUE)
extended_scores <- democracy_scores(extended_model)
extended_scores <- bind_cols(data,extended_scores)
# Matches the calculated scores to the level of the 2014 release of the UD scores
extended_scores <- match_to_uds(extended_scores)}
}

