% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickUDS.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare democracy data before replicating the UDS model}
\usage{
prepare_data(data, .funs)
}
\arguments{
\item{data}{A dataset of democracy scores. For the function to do anything,
the column names must contain at least one of the following strings:
\code{arat}, \code{blm}, \code{bmr}, \code{bollen}, \code{doorenspleet},
\code{wgi}, \code{gwf}, \code{hadenius}, \code{kailitz}, \code{lied},
\code{munck}, \code{pacl}, \code{peps}, \code{polyarchy}, \code{polity},
\code{prc}, \code{PIPE}, \code{svolik}, \code{ulfelder}, \code{utip},
\code{v2x}, \code{vanhanen_democratization}, \code{vanhanen_pmm}, or
\code{wth}. For details of these variables, see the documentation for
\code{\link{democracy}}.}

\item{.funs}{A names list of functions to modify the columns. It defaults to
the following:

\code{funs(arat = cut(., breaks = c(0, 50, 60, 70, 80, 90, 100, 109),
  labels = 1:7, include.lowest = TRUE, right = FALSE),

  hadenius = cut(., breaks =  c(0, 1, 2, 3, 4, 7, 8, 9, 10), labels = 1:8,
  include.lowest = TRUE, right = FALSE),

  bollen = cut(., breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, 90, 100),
  labels = 1:10, include.lowest = TRUE, right = FALSE),


  vanhanen = cut(., breaks = c(0, 5, 10, 15, 20, 25, 30, 35, 50), labels =
  1:8, include.lowest = TRUE, right = FALSE),

  munck = cut(., breaks = c(0, 0.5, 0.75, 0.99, 1), labels = 1:4,
  include.lowest = TRUE, right = FALSE),


  polyarchy_dimensions = cut(., breaks = 20, include.lowest = TRUE, right =
  FALSE, ordered_result = TRUE),

  polity = ifelse(. < -10, NA, .), v2x = cut(., breaks = 20, include.lowest =
  TRUE, right = FALSE, ordered_result = TRUE),

  wgi = cut(., breaks = 20, include.lowest = TRUE, right = FALSE,
  ordered_result = TRUE),

  peps = round(.),

  other = as.numeric(unclass(factor(.))))}}
}
\value{
A data frame with the transformed scores, if any.
}
\description{
This function is designed to take the democracy data included in this package
and put it in a form suitable for use with the \code{\link{mirt}} package to
replicate the UDS model. It takes a data frame and tries to determine, from
the colum names, which variables contain democracy scores.
}
\details{
If the column names contain the strings \code{arat}, \code{blm},
\code{bollen},\code{wgi}, \code{hadenius}, \code{munck}, \code{pacl},
\code{peps}, \code{polyarchy_inclusion_dimension},
\code{polyarchy_contestation_dimension}, \code{polity}, \code{prc},
\code{v2x}, \code{vanhanen_pmm}, or \code{vanhanen_democratization}, the
function performs the following transformations by default:

\code{arat}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function cuts Arat (1991)'s 0-109
democracy score into 7 intervals with the following cutoffs: 50, 60, 70, 80,
90, and 100. The resulting score is ordinal from 1 to 8.

\code{bollen}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function cuts Bollen's (2001)'s
0-100 democracy score into 10 intervals with the following cutoffs:
10,20,30,40,50,60,70,80, and 90. The resulting score is ordinal from 1 to 10.

\code{wgi}: If the World Governance Indicator's index of voice and
acocuntability is included in the file, the function cuts it into 20
categories. The resulting score is ordinal from 1 to 20.

\code{hadenius}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function cuts Hadenius (1992)'s
0-10 democracy score into 8 intervals with the following cutoffs: 1, 2,3,4,
7, 8, and 9. The resulting score is ordinal from 1 to 8.

\code{munck}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function cuts Munck's (2009)'s 0-1
democracy score into 4 intervals with the following cutoffs: 0.5,0.5,0.75,
and 0.99. The resulting score is ordinal from 1 to 4.

\code{peps}: If any of the variants of the Participation-Enhanced Polity
Score (Moon et al 2006) is included in the file, the function rounds its
value (eliminates the decimal) and then transforms it into an ordinal measure
from 1 to 21.

\code{polity}: Following Pemstein, Meserve, and Melton's replication code
(Pemstein, Meserve, and Melton 2013), the function takes the polity scores
and puts NA for any values below -10, and then transforms it into an ordinal
measure from 1 to 21.

\code{polyarchy_inclusion_dimension},
\code{polyarchy_contestation_dimension}: If any of the polyarchy inclusion or
contestation dimensions from Coppedge, Alvarez and Maldonado (2008)  are
included, it cuts them into into 20 categories. The resulting score is
ordinal from 1 to 20.

\code{v2x}: If any of the v2x_ continuous indexes of democracy from the V-Dem
dataset (Coppedge et al 2015) are included in the file, the function cuts
them into 20 categories. The resulting score is ordinal from 1 to 20.

\code{vanhanen_democratization} or \code{vanhanen_pmm}: Following Pemstein,
Meserve, and Melton's replication code (Pemstein, Meserve, and Melton 2013),
the function cuts Vanhanen's (2012)'s index of democratization into 8
intervals with the following cutoffs: 5,10,15,20,25,30, and 35. The resulting
score is ordinal from 1 to 8.

The function also recognizes the following column names (or partial column
names - it also recognizes, e.g., pmm_blm) as measures of democracy:
\code{blm} (from Bowman, Lehoucq, and Mahoney 2005), \code{bmr} (from Boix,
Miller, and Rosato 2012), \code{doorenspleet} (from Doorenspleet 2000),
\code{e_v2x} (the "ordinal" indexes from the V-dem project, Coppedge at al
2015), \code{freedomhouse} or \code{fh} (from Freedom House - freedom scale
must be reversed so that "more freedom" is higher), \code{gwf} (from Geddes,
Wright, and Frantz 2014 - the dichotomous democracy indicator only),
\code{kailitz} (from Kailitz 2013 - democracy/non-democracy indicator),
\code{lied} or \code{lexical_index} (from Skaaning, Gerring, and
Bartusevicius 2015), \code{mainwaring} (from Mainwaring and Perez Linan
2008), \code{magaloni} (from Magaloni, Min, Chu 2013 -
democracy/non-democracy indicator), \code{pacl} (from Cheibub, Gandhi, and
Vreeland 2010), \code{pitf} (from Goldstone et al 2010 or Taylor and Ulfelder
2015), \code{polyarchy} (from Coppedge and Reinicke 1991), \code{prc} (from
Gasiorowski 1996 or Reich 2002), \code{PIPE} (from Przeworski 2010),
\code{reign} (from Bell 2016), \code{svolik} (from Svolik 2012,
democracy/dictatorship indicator only), \code{ulfelder} (from Ulfelder 2012),
\code{utip} (from Hsu 2008), and \code{wth} or \code{wahman_teorell_hadenius}
(from Wahman, Teorell, and Hadenius 2013). In each of these cases the
function transforms the values of these scores by running
\code{as.numeric(unclass(factor(x)))}, which transforms them into ordinal
variables from 1 to the number of categories.

For details of these scores, see the documentation for
\code{\link{democracy}}.

It is also possible to change these defaults.
}
\section{Note}{
 Warning! The function does not perform any sanity checks. It
will try to transform anything that has the right name. You should always
check the results make sense.
}

\examples{
summary(democracy)
summary(prepare_data(democracy))
}
\references{
Arat, Zehra F. 1991. Democracy and human rights in developing countries.
Boulder: Lynne Rienner Publishers.

Bell, Curtis. 2016. The Rulers, Elections, and Irregular Governance Dataset
(REIGN).\url{http://oefresearch.org/datasets/reign}

Boix, Carles, Michael Miller, and Sebastian Rosato. 2012. A Complete Data Set
of Political Regimes, 1800-2007. Comparative Political Studies 46 (12):
1523-1554. Original data available at
\url{https://sites.google.com/site/mkmtwo/democracy-v2.0.dta?attredirects=0}

Bollen, Kenneth A. 2001. "Cross-National Indicators of Liberal Democracy,
1950-1990." 2nd ICPSR version. Chapel Hill, NC: University of North Carolina,
1998. Ann Arbor, MI: Inter-university Consortium for Political and Social
Research, 2001. Original data available at
\url{http://webapp.icpsr.umich.edu/cocoon/ICPSR-STUDY/02532.xml}.

Bowman, Kirk, Fabrice Lehoucq, and James Mahoney. 2005. Measuring Political
Democracy: Case Expertise, Data Adequacy, and Central America. Comparative
Political Studies 38 (8): 939-970.
\url{http://cps.sagepub.com/content/38/8/939}. Data available at
\url{http://www.blmdemocracy.gatech.edu/}.

Cheibub, Jose Antonio, Jennifer Gandhi, and James Raymond Vreeland. 2010.
"Democracy and Dictatorship Revisited." Public Choice. 143(1):67-101.
Original data available at
\url{https://sites.google.com/site/joseantoniocheibub/datasets/democracy-and-dictatorship-revisited}.

Coppedge, Michael, John Gerring, Staffan I. Lindberg, Svend-Erik Skaaning,
and Jan Teorell, with David Altman, Michael Bernhard, M. Steven Fish, Adam
Glynn, Allen Hicken, Carl Henrik Knutsen, Kelly McMann, Pamela Paxton, Daniel
Pemstein, Jeffrey Staton, Brigitte Zimmerman, Frida Andersson, Valeriya
Mechkova, Farhad Miri. 2015. V-Dem Codebook v5. Varieties of Democracy
(V-Dem) Project. Original data available at \url{https://v-dem.net/en/data/}.

Coppedge, Michael and Wolfgang H. Reinicke. 1991. Measuring Polyarchy. In On
Measuring Democracy: Its Consequences and Concomitants, ed. Alex Inkeles. New
Brunswuck, NJ: Transaction pp. 47-68.

Coppedge, A. Alvarez and C. Maldonado. 2008. "Two Persistent Dimensions of
Democracy: Contestation and Inclusiveness". The Journal of Politics 70.03,
pp. 632-647. DOI: 10.1017/S0022381608080663.

Doorenspleet, Renske. 2000. Reassessing the Three Waves of Democratization.
World Politics 52 (03): 384-406. DOI: 10.1017/S0043887100016580.
\url{http://dx.doi.org/10.1017/S0043887100016580}.

Kaufmann, D. and A. Kraay. 2016. Worldwide Governance Indicators. 2016.
\url{http://www.govindicators.org}.

Freedom House. 2015. "Freedom in the World." Original data available at
\url{http://www.freedomhouse.org}.

Gasiorowski, Mark J. 1996. "An Overview of the Political Regime Change
Dataset." Comparative Political Studies 29(4):469-483.

Geddes, Barbara, Joseph Wright, and Erica Frantz. 2014. Autocratic Breakdown
and Regime Transitions: A New Data Set. Perspectives on Politics 12 (1):
313-331. Original data available at \url{http://dictators.la.psu.edu/}.

Goldstone, Jack, Robert Bates, David Epstein, Ted Gurr, Michael Lustik, Monty
Marshall, Jay Ulfelder, and Mark Woodward. 2010. A Global Model for
Forecasting Political Instability. American Journal of Political Science 54
(1): 190-208. DOI:10.1111/j.1540-5907.2009.00426.x

Hadenius, Axel. 1992. Democracy and Development. Cambridge: Cambridge
University Press.

Hadenius, Axel & Jan Teorell. 2007. "Pathways from Authoritarianism", Journal
of Democracy 18(1): 143-156.

Hsu, Sara "The Effect of Political Regimes on Inequality, 1963-2002," UTIP
Working Paper No. 53 (2008), http://utip.gov.utexas.edu/papers/utip_53.pdf.
Data available for download at http://utip.gov.utexas.edu/data/.

Kailitz, Steffen. 2013. Classifying political regimes revisited: legitimation
and durability. Democratization 20 (1): 39-60. Original data available at
\url{http://dx.doi.org/10.1080/13510347.2013.738861}.

Mainwaring, Scott, Daniel Brinks, and Anibal Perez Linan. 2008. "Political
Regimes in Latin America, 1900-2007." Original data available from
\url{http://kellogg.nd.edu/scottmainwaring/Political_Regimes.pdf}.

Magaloni, Beatriz, Jonathan Chu, and Eric Min. 2013. Autocracies of the
World, 1950-2012 (Version 1.0). Dataset, Stanford University. Original data
and codebook available at
\url{http://cddrl.fsi.stanford.edu/research/autocracies_of_the_world_dataset/}

Marshall, Monty G., Ted Robert Gurr, and Keith Jaggers. 2012. "Polity IV:
Political Regime Characteristics and Transitions, 1800-2012." Updated to
2015. Original data available from
\url{http://www.systemicpeace.org/polity/polity4.htm}.

Moon, Bruce E., Jennifer Harvey Birdsall, Sylvia Ceisluk, Lauren M. Garlett,
Joshua J. Hermias, Elizabeth Mendenhall, Patrick D. Schmid, and Wai Hong Wong
(2006) "Voting Counts: Participation in the Measurement of Democracy" Studies
in Comparative International Development 42, 2 (Summer, 2006). The complete
dataset is available here:
\url{http://www.lehigh.edu/~bm05/democracy/Obtain_data.htm}.

Munck, Gerardo L. 2009. Measuring Democracy: A Bridge Between Scholarship and
Politics. Baltimore: Johns Hopkins University Press.

Pemstein, Daniel, Stephen Meserve, and James Melton. 2010. Democratic
Compromise: A Latent Variable Analysis of Ten Measures of Regime Type.
Political Analysis 18 (4): 426-449.

Pemstein, Daniel, Stephen A. Meserve, and James Melton. 2013. "Replication
data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures
of Regime Type." In: Harvard Dataverse.
\url{http://hdl.handle.net/1902.1/PMM}

Przeworski, Adam et al. 2013. Political Institutions and Political Events
(PIPE) Data Set. Department of Politics, New York University.
\url{https://sites.google.com/a/nyu.edu/adam-przeworski/home/data}

Reich, G. 2002. Categorizing Political Regimes: New Data for Old Problems.
Democratization 9 (4): 1-24.
\url{http://www.tandfonline.com/doi/pdf/10.1080/714000289}.

Skaaning, Svend-Erik, John Gerring, and Henrikas Bartusevicius. 2015. A
Lexical Index of Electoral Democracy. Comparative Political Studies 48 (12):
1491-1525. Original data available from
\url{http://thedata.harvard.edu/dvn/dv/skaaning}.

Svolik, Milan. 2012. The Politics of Authoritarian Rule. Cambridge and New
York: Cambridge University Press. Original data available from
\url{http://campuspress.yale.edu/svolik/the-politics-of-authoritarian-rule/}.

Taylor, Sean J. and Ulfelder, Jay, A Measurement Error Model of Dichotomous
Democracy Status (May 20, 2015). Available at SSRN:
\url{http://ssrn.com/abstract=2726962} or
\url{http://dx.doi.org/10.2139/ssrn.2726962}

Ulfelder, Jay. 2012. "Democracy/Autocracy Data Set." In: Harvard Dataverse.
\url{http://hdl.handle.net/1902.1/18836}.

Vanhanen, Tatu. 2012. "FSD1289 Measures of Democracy 1810-2012." Original
data available from
\url{http://www.fsd.uta.fi/english/data/catalogue/FSD1289/meF1289e.html}

Wahman, Michael, Jan Teorell, and Axel Hadenius. 2013. Authoritarian regime
types revisited: updated data in comparative perspective. Contemporary
Politics 19 (1): 19-34.
}
