% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickUDS.R
\name{prepare_democracy}
\alias{prepare_democracy}
\title{Prepares selected indexes in the democracy dataset for use in a UD model}
\usage{
prepare_democracy(..., identifiers = c("extended_country_name", "year", "GWn",
  "cown", "polity_ccode", "in_GW_system"))
}
\arguments{
\item{...}{a set of measures of democracy in the \code{\link{democracy}}
dataset (either bare column names, or strings). Can be a \link[dplyr:select]{dplyr::select}
expression (but \link{dplyr} must be loaded for this to work).}

\item{identifiers}{A set of identifiers to return with the data. Defaults to
\code{c("extended_country_name", "year", "GWn", "cown", "polity_ccode", "in_GW_system")}.
Must exist in the \link{democracy} dataset.}
}
\value{
A dataset ready for use with \code{\link{democracy_model}}
}
\description{
Prepares selected indexes in the democracy dataset for use in a UD model
}
\examples{
prepare_democracy(dplyr::matches("pmm"))
}
