% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick UDS.R
\name{raterinfo}
\alias{raterinfo}
\title{A convenience function for extracting rater info from a UD model in a tidy
data frame format.}
\usage{
raterinfo(model)
}
\arguments{
\item{model}{A \code{\link{mirt}} \code{\link{SingleGroupClass-class}} model
of the democracy scores.}
}
\value{
A data frame with rater information for each democracy index over the
  range of the latent variable \code{theta}.
}
\description{
A convenience function for extracting rater info from a UD model in a tidy
data frame format.
}
\examples{
# Not run:
# library(QuickUDS)
# data <- prepare_data(democracy)
# data <- melt(data, measure.vars = names(data)[grep("pmm",names(data))], na.rm = TRUE)
# data <- data \%>\% group_by(country_name,year) \%>\% mutate(num_measures = n())
# data <- dcast(data, ... ~ variable)
# data <- data \%>\% arrange(country_name,year)
# replication_2011_model <- mirt(data[ , names(data)[grep("pmm",names(data))],
#                                model = 1, itemtype = "graded", SE = TRUE)
# raterinfo_2011 <- raterinfo(replication_2011_model)
# head(raterinfo)
}

