% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickUDS.R
\name{democracy_model}
\alias{democracy_model}
\title{Produce a UD model from democracy data}
\usage{
democracy_model(data, columns, model = 1, itemtype = "graded", SE = TRUE,
  ...)
}
\arguments{
\item{data}{The prepared democracy data. Run \code{\link{prepare_data}} first
on your democracy data. Note that \code{\link{mirt}} will throw an error if
some country-years are not measured by any dataset; it is thus necessary to ensure
that your data has at least one measurement of democracy for every
country-year (though NAs are fine in many of the variables). See the
examples for details.}

\item{columns}{The columns to use to estimate the model.}

\item{model}{The model type for \code{\link{mirt}}. The default, 1,
calculates a one-factor model (a single latent variable).}

\item{itemtype}{The type of item to estimate. See \code{\link{mirt}} for
details. The default fits a "graded" model like the one Pemstein, Meserve,
and Melton use in their 2010 paper.}

\item{SE}{Boolean. Whether to calculate the standard errors of the model
parameters. The default is \code{TRUE}.}

\item{...}{Other parameters passed on to \code{\link{mirt}}}
}
\value{
a \code{\link{SingleGroupClass-class}} model of latent democracy
  scores suitable for use by \code{\link{democracy_scores}}.
}
\description{
This function is a simple wrapper for \code{mirt(data[, columns], model = 1,
itemtype = "graded", SE = TRUE, ...)}. More fine-grained control can be
achieved by using \code{\link{mirt}} directly.
}
\examples{
\donttest{
data <- prepare_data(democracy)
data <- melt(data, measure.vars = names(data)[grep("pmm",names(data))], na.rm = TRUE)
data <- data \%>\% group_by(country_name,year) \%>\% mutate(num_measures = n())
data <- dcast(data, ... ~ variable)
data <- data \%>\% arrange(country_name,year)
replication_2011_model <- democracy_model(data,
                               columns = names(data)[grep("pmm",names(data))])}
}

