% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quick UDS.R
\name{prepare_democracy}
\alias{prepare_democracy}
\title{Prepares the selected indexes in the \code{\link{democracy}} dataset}
\usage{
prepare_democracy(indexes)
}
\arguments{
\item{indexes}{a set of measures of democracy in the \code{\link{democracy}}
dataset. They must be names of the columns of the \code{\link{democracy}}
dataset - otherwise the function throws an error.}
}
\value{
A dataset ready for use with \code{\link{democracy_model}}
}
\description{
Prepares the selected indexes in the \code{\link{democracy}} dataset
}
\examples{
prepare_democracy(names(democracy)[grep("pmm",names(democracy))])
}

