% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuickUDS.R
\name{raterinfo}
\alias{raterinfo}
\title{Extract rater info from a UD model in a tidy format.}
\usage{
raterinfo(model)
}
\arguments{
\item{model}{A \code{\link{mirt}} \code{\link{SingleGroupClass-class}} model
of the democracy scores.}
}
\value{
A data frame with rater information for each democracy index over the
range of the latent variable \code{theta}.
}
\description{
Extract rater info from a UD model in a tidy format.
}
\examples{
\donttest{
replication_2011_model <- democracy_model(dplyr::matches("pmm"), verbose = FALSE)
raterinfo(replication_2011_model)}
}
